All data is now extracted successfuly #49
